+++ projects/hipsparse/library/src/nvidia_detail/hipsparse_generic_auxiliary.cpp
@@ -494,6 +494,98 @@ hipsparseStatus_t hipsparseCreateConstSlicedEll(hipsparseConstSpMatDescr_t* spMa
 }
 #endif
 
+
+
+
+
+
+
+
+
+
+#if(CUDART_VERSION >= 12011)
+hipsparseStatus_t hipsparseCreateBsr(hipsparseSpMatDescr_t* spMatDescr,
+                                     int64_t                mb,
+                                     int64_t                nb,
+                                     int64_t                nnzb,
+                                     hipsparseDirection_t   blockDir,
+                                     int64_t                blockDim,
+                                     void*                  bsrRowPtr,
+                                     void*                  bsrColInd,
+                                     void*                  bsrValues,
+                                     hipsparseIndexType_t   bsrRowPtrType,
+                                     hipsparseIndexType_t   bsrColIndType,
+                                     hipsparseIndexBase_t   idxBase,
+                                     hipDataType            valueType)
+{
+    return hipsparse::hipCUSPARSEStatusToHIPStatus(
+        cusparseCreateBsr((cusparseSpMatDescr_t*)spMatDescr,
+                                     mb,
+                                     nb,
+                                     nnzb,
+                                     hipsparse::hipDirectionToCudaDirection(blockDir),
+                                     blockDim,
+                                     bsrRowPtr,
+                                     bsrColInd,
+                                     bsrValues,
+                                     hipsparse::hipIndexTypeToCudaIndexType(bsrRowPtrType),
+                                     hipsparse::hipIndexTypeToCudaIndexType(bsrColIndType),
+                                     hipsparse::hipIndexBaseToCudaIndexBase(idxBase),
+                                     hipsparse::hipDataTypeToCudaDataType(valueType)));
+}
+#endif
+
+#if(CUDART_VERSION >= 12011)
+hipsparseStatus_t hipsparseCreateConstBsr(hipsparseConstSpMatDescr_t* spMatDescr,
+                                          int64_t                     mb,
+                                          int64_t                     nb,
+                                          int64_t                     nnzb,
+                                          hipsparseDirection_t        blockDir,
+                                          int64_t                     blockDim,
+                                          const void*                 bsrRowPtr,
+                                          const void*                 bsrColInd,
+                                          const void*                 bsrValues,
+                                          hipsparseIndexType_t        bsrRowPtrType,
+                                          hipsparseIndexType_t        bsrColIndType,
+                                          hipsparseIndexBase_t        idxBase,
+                                          hipDataType                 valueType)
+{
+    return hipsparse::hipCUSPARSEStatusToHIPStatus(
+        cusparseCreateConstBsr((cusparseConstSpMatDescr_t*)spMatDescr,
+                                     mb,
+                                     nb,
+                                     nnzb,
+                                     hipsparse::hipDirectionToCudaDirection(blockDir),
+                                     blockDim,
+                                     bsrRowPtr,
+                                     bsrColInd,
+                                     bsrValues,
+                                     hipsparse::hipIndexTypeToCudaIndexType(bsrRowPtrType),
+                                     hipsparse::hipIndexTypeToCudaIndexType(bsrColIndType),
+                                     hipsparse::hipIndexBaseToCudaIndexBase(idxBase),
+                                     hipsparse::hipDataTypeToCudaDataType(valueType)));
+}
+#endif
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 #if(CUDART_VERSION >= 12000)
 hipsparseStatus_t hipsparseDestroySpMat(hipsparseConstSpMatDescr_t spMatDescr)
 {