+++ projects/hipsparse/CHANGELOG.md
@@ -7,7 +7,7 @@ Documentation for hipSPARSE is available at
 ## Since last release ROCm 7.13
 
 ### Added
-* Add `hipsparseCreateBsr` in order to enable BSR format support in generic routines.
+* Add `hipsparseCreateBsr` and `hipsparseCreateConstBsr` in order to enable BSR format support in generic routines.
 * Add BSR format support to `hipsparseSpMV`.
 
 ## hipSPARSE 4.5.0 for ROCm 7.12.0